test(vetting): Wave 2 — migrate GLCM tests to the taxonomy (rename-in-place)
Rename the 5 live GLCM test files to the SPEC naming and update test_all.cc includes:
  test_glcm.h            -> test_glcm_regression.h
  test_glcm_oracle.h     -> test_glcm_mechanics.h   (it was a mechanics guard, not an oracle)
  test_ibsi_glcm.h       -> test_glcm_ibsi.h
  test_compat_3d_glcm.h  -> test_3d_glcm_pyradiomics.h
  test_glcm_oracle.py    -> test_glcm_pyradiomics.py (the real 2D pyradiomics oracle)

Pure renames (git mv, history preserved); no test content changed. Verified locally
in a conda build: full gtest suite 696/696 pass, GLCM 166/166. test_3d_glcm.h left
untouched and flagged orphaned in the registry (not #included; 26 fns never run).
Registry current_test repointed; MIGRATION.md 5.9 records the map corrections.
