+++ tests/test_all.cc
@@ -15,7 +15,7 @@
 #include "test_2d_remaining_features.h"
 #include "test_neighbors_2d.h"
 #include "test_initialization.h"
-#include "test_ibsi_glcm.h"
+#include "test_glcm_ibsi.h"
 #include "test_ibsi_gldm.h"
 #include "test_ibsi_glrlm.h"
 #include "test_ibsi_gldzm.h"
@@ -23,7 +23,7 @@
 #include "test_ibsi_intensity.h"
 #include "test_ibsi_ngldm.h"
 #include "test_ibsi_ngtdm.h"
-#include "test_glcm.h"
+#include "test_glcm_regression.h"
 #include "test_gldm.h"
 #include "test_gldm_oracle.h"
 #include "test_glrlm.h"
@@ -38,13 +38,13 @@
 #include "test_3d_gldzm.h"
 #include "test_3d_ngldm.h"
 #include "test_compat_3d_fo_radiomics.h"
-#include "test_compat_3d_glcm.h"
+#include "test_3d_glcm_pyradiomics.h"
 #include "test_compat_3d_gldm.h"
 #include "test_compat_3d_ngtdm.h"
 #include "test_compat_3d_glrlm.h"
 #include "test_compat_3d_glszm.h"
 #include "test_3d_feature_coverage.h"
-#include "test_glcm_oracle.h"
+#include "test_glcm_mechanics.h"
 #ifdef USE_ARROW
     #include "test_arrow.h"
     #include "test_arrow_file_name.h"
