+++ tests/test_glcm_mechanics.h
